Cosmetic improvements

+++ bgra2yuv420p.h
@@ -24,8 +24,8 @@
 #include <stddef.h>
 #include <stdint.h>
 
-void bgra2yuv420p(const uint8_t *restrict bgra, int bgra_stride,
-	uint8_t *restrict y, int y_stride, uint8_t *restrict u, int u_stride,
-	uint8_t *restrict v, int v_stride, size_t width, size_t height);
+void bgra2yuv420p(const uint8_t *restrict bgra, ptrdiff_t bgra_stride,
+	uint8_t *restrict y, ptrdiff_t y_stride, uint8_t *restrict u, ptrdiff_t u_stride,
+	uint8_t *restrict v, ptrdiff_t v_stride, size_t width, size_t height);
 
 #endif
